refactor(game): extract frontier candidate pickers to lib/game/frontier
Moves the play-phase explore candidate selection out of the 7,865-line
data-server.ts into lib/game/frontier.ts:
  - pickFrontierCandidate (ring-walk vs Monte-Carlo two-phase picker)
  - pickByRingWalk / buildFrontierSample / countHostileNeighbors (helpers)
  - FRONTIER_MAX_RINGS / EXPLORE_MONTE_CARLO_THRESHOLD / MONTE_CARLO_MAX_SAMPLES

These are dependency-injected (take `db`) and hold no transaction/write state,
so they extract cleanly: frontier.ts imports only from ./world-gen,
./exploration and ./data-access/collections — no data-server import, no cycle.
`frontierExploreServer` (the transactional orchestrator) and
`pickFrontierCandidatesBulk` stay in data-server and import what they need
(pickFrontierCandidate + the three constants).

Behavior-preserving — code moved verbatim. data-server.ts 7,865 → 7,660 lines.

Verified: game suite 1460/1460, explore/frontier app tests 33/33, tsc + eslint clean.

Stacked on the Phase-4 foundation PR (#1567); retargets to develop once that merges.
